1、升级springboot版本，引入springboot AI 2、新增Repository，以及es和pg的实现

+++ ytx-agent-framework-llm/src/main/java/com/ytx/ai/agent/llm/service/LlmService.java
@@ -4,7 +4,6 @@ import com.plexpt.chatgpt.entity.chat.ChatCompletion;
 import com.plexpt.chatgpt.entity.chat.ChatCompletionResponse;
 import com.ytx.ai.agent.llm.vo.LlmChatCompletion;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public interface LlmService {
@@ -14,5 +13,5 @@ public interface LlmService {
     public String chatCompletion(LlmChatCompletion llmChatCompletion);
 
 
-    public List<BigDecimal> createEmbeddings(String content);
+    public List<Float> createEmbeddings(String content);
 }
